requirements.txt
pandas==2.2.1
numpy==1.26.0
biopython==1.83